requirements.txt
numpy~=1.22.4
matplotlib~=3.5.2
Pillow~=9.1.1
torch~=1.11.0+cu113
torchvision~=0.12.0+cu113
einops~=0.6.0
timm~=0.6.12
rich~=12.5.1
